+++ packages/react-components/src/button/components/iconWithTextButton.js
@@ -4,7 +4,7 @@ import styled from 'styled-components'
 // utils
 import { getIconWithTextButtonTheme } from '../utils/theme'
 // components
-import { P4 } from '../../text/paragraph'
+import { P3 } from '../../text/paragraph'
 // @twreporter
 import mq from '@twreporter/core/lib/utils/media-query'
 import { THEME } from '@twreporter/core/lib/constants/theme'
@@ -14,26 +14,27 @@ const ButtonContainer = styled.div`
   display: flex;
   flex-direction: column;
   align-items: center;
-  color: ${props => props.color};
+  color: ${(props) => props.color};
   svg {
     width: 24px;
     height: 24px;
-    background-color: ${props => props.color};
+    background-color: ${(props) => props.color};
   }
 
   ${mq.desktopAndAbove`
     &:hover {
-      color: ${props => props.hoverColor};
+      color: ${(props) => props.hoverColor};
       svg {
-        background-color: ${props => props.hoverColor};
+        background-color: ${(props) => props.hoverColor};
       }
     }
   `}
 `
 
-const StyledP4 = styled(P4)`
-  max-height: ${props => (props.hideText ? '0px' : 'none')};
-  opacity: ${props => (props.hideText ? '0' : '1')};
+const StyledP3 = styled(P3)`
+  margin-top: 2px;
+  max-height: ${(props) => (props.hideText ? '0px' : 'none')};
+  opacity: ${(props) => (props.hideText ? '0' : '1')};
   transition: opacity 100ms;
 `
 
@@ -54,7 +55,7 @@ const IconWithTextButton = ({
   return (
     <ButtonContainer color={color} hoverColor={hoverColor}>
       {iconComponent}
-      <StyledP4 text={text} weight="bold" hideText={hideText} />
+      <StyledP3 text={text} weight={P3.Weight.NORMAL} hideText={hideText} />
     </ButtonContainer>
   )
 }